Move the agentic integration class to the openai platform
